+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleases](https://github.com/quartiq/miniconf/compare/v0.17.2...HEAD) - DATE
+
+### Changed
+
+* `KeyLookup`, `Walk`, and `Metadata` layout and behavior
+
 ## [0.17.2](https://github.com/quartiq/miniconf/compare/v0.17.1...v0.17.2) - 2024-11-19
 
 ### Added